final changes to branch before merge trixies to master

+++ bin/base-rebuild.sh
@@ -175,16 +175,20 @@ cd $BUILDHOME
 #DOCKERFILE="Dockerfile"
 # Because .... rust ... I don't know why .... 
 # We build the rust crates here so we can pass them into the build container. 
-echo "Building openvas_scanner rust crates"
-rm -rf /build/*
-cd /build
-wget --no-verbose https://github.com/greenbone/openvas-scanner/archive/$openvas.tar.gz
-tar -zxf $openvas.tar.gz
-cd /build/*/
-cd rust
-make
-tar cvf crates.tar crates
-mv crates.tar $BUILDHOME/rust
+# We if then this, because if we don't, it forces the buildx to rebuild everything.
+if ! [ -f $BUILDHOME/rust/$openvas ]; then
+	echo "Building openvas_scanner rust crates"
+	rm -rf /build/*
+	cd /build
+	wget --no-verbose https://github.com/greenbone/openvas-scanner/archive/$openvas.tar.gz
+	tar -zxf $openvas.tar.gz
+	cd /build/*/
+	cd rust
+	make
+	tar cvf $openvas.crates.tar crates
+	mv $openvas.crates.tar $BUILDHOME/rust/crates.tar
+	touch $BUILDHOME/rust/$openvas 
+fi
 cd $BUILDHOME
 sync
 echo "Working in $(pwd)"
@@ -192,7 +196,6 @@ ls -l rust
 # Now build everything together. At this point, this will normally only be the arm7 build as the amd64 was likely built and cached as beta.
 SLIMSTART=$(date +%s)
 docker buildx build $PUBLISH \
-   --progress=plain \
    --platform $arch -f Dockerfile \
    --target slim \
    -t immauss/openvas:${tag}-slim \
@@ -203,7 +206,6 @@ SLIMFIN=$(date +%s)
 
 FINALSTART=$(date +%s)
 docker buildx build $PUBLISH \
-	--progress=plain \
 	--platform $arch \
    	--target final \
 	-t immauss/openvas:${tag} \
